feature_flags_backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
gunicorn>=21.2
psycopg2-binary>=2.9
pytest>=7.0
pytest-django>=4.5 